requirements.txt
# fastapi
# uvicorn
# pillow
# opencv-python-headless
# torch
# torchvision
# python-multipart
fastapi
uvicorn
python-multipart
fpdf
pillow
opencv-python
